better burn ins, comments
choose more burn ins for the mcmc part to make sure we wind up in the
right high confidence region, then do the ‘mcmc for real’ and beat up
on that region. also added some plots of the model & data for part 1
and redid the corner plots for the burn in results and ‘for real’ mcmc
results. calculated confidence regions